requirements.txt
fastapi
arq
numpy
tensorstore
uvicorn
fastapi
pillow
redis
requests